refactor: remove sourcemap dir
