+++ bundler/CMakeLists.txt
@@ -18,8 +18,8 @@ set(SOURCE_FILES
         src/GlobalImportHandler.cpp
         src/Error.h
         src/Error.cpp
-        src/sourcemap/SourceMapGenerator.h
-        src/sourcemap/SourceMapGenerator.cpp
+        src/SourceMapGenerator.h
+        src/SourceMapGenerator.cpp
         src/codegen/NodeTraverser.h
         src/codegen/NodeTraverser.cpp
         src/codegen/CodeGen.h
