+++ bundler/src/SourceMapGenerator.cpp
