+++ bundler/src/SourceMapGenerator.h
